+++ Carro/Carro/Carro/ViewModels/EditarDespesaViewModel.cs
@@ -91,8 +91,8 @@ namespace Carro.ViewModels
             }
         }
 
-        float _valorEntry = 0.0F;
-        public float valorEntry
+        decimal _valorEntry = 0.0m;
+        public decimal valorEntry
         {
             get
             {
